requirements.txt
PyYAML==6.0.1
kubernetes==28.1.0
click==8.1.7
